bin/ci: pass POSTGRES_HOST_AUTH_METHOD=trust to postgres container
In https://github.com/docker-library/postgres/pull/658 a change was
made to the postgres image to require a password by default or
explicit disabling of auth.

For bin/ci, disable auth to keep configs simple. Add a note reminding
that doing this is highly insecure and should not be done in production.